Restructure directory to expose store instance
