+++ webpack.config.js
@@ -3,7 +3,7 @@
 const join = require('path').join;
 
 module.exports = {
-  entry: './source/client/js/index.js',
+  entry: './source/client/index.js',
 
   output: {
     filename: 'bundle.js',
